Refactor tail_n
Go to the end of the file and read backwards one character at a time checking if it's a line break. Continue until n+1 line breaks are found OR if reached the beginning of the file. Then print out the contents.

This change fixes printing an additional line break when the number of lines requested is less than the total number of lines contained in the file. It also fixes the infinite loop when asking for more lines than the file contains.