Rewrite tail_n without side-effects
The purpose of this rewrite is to have tail_n be predictable, reliable and testable. An orchestrator will call tail_n. In return tail_n will set the file stream to the position requested and return without side-effects. The caller will then do what it needs to with the file stream.