Configure parse_arguments to be re-entrant
From http://man7.org/linux/man-pages/man3/getopt.3.html,

"The variable optind is the index of the next element to be processed
in argv.  The system initializes this value to 1.  The caller can
reset it to 1 to restart scanning of the same argv, or when scanning
a new argument vector."